// coc-dan-frontend/src/bindings/api/ws/tx/ITxEvent.d.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ITxAction } from "../../../entity/tx/ITxAction.d";

export type ITxEvent = { tx_id: bigint, raw_id: bigint, stage_id: bigint, user_id: bigint, avatar_id: string, time: string, action: ITxAction, };